Factor out the synchronisation of changed quota on a per-storage basis
