+++ lib/vsc/administration/user.py
@@ -33,10 +33,12 @@ from urllib2 import HTTPError
 
 from vsc.utils import fancylogger
 from vsc.accountpage.wrappers import VscAccount, VscAccountPerson, VscAccountPubkey, VscHomeOnScratch, VscUserGroup
+from vsc.accountpage.wrappers import mkVscAccount, mkUserGroup
 from vsc.accountpage.wrappers import VscGroup, VscUserSizeQuota
 from vsc.administration.institute import Institute
 from vsc.administration.tools import create_stat_directory
 from vsc.config.base import VSC, Muk, VscStorage, VSC_DATA, VSC_HOME
+from vsc.config.base import NEW, MODIFIED, MODIFY, ACTIVE
 from vsc.filesystem.ext import ExtOperations
 from vsc.filesystem.gpfs import GpfsOperations
 from vsc.filesystem.posix import PosixOperations
@@ -908,3 +910,116 @@ cluster_user_pickle_store_map = {
     'delcatty': 'VSC_SCRATCH_DELCATTY',
     'muk': 'VSC_SCRATCH_MUK',
 }
+
+def update_user_status(user, options, client):
+    """
+    Change the status of the user's account in the account page to active. Do the same for the corresponding UserGroup.
+    """
+    if user.dry_run:
+        logger.info("User %s has account status %s. Dry-run, not changing anything", user.user_id, user.account.status)
+        return
+
+    if user.account.status not in (NEW, MODIFIED, MODIFY):
+        logger.info("Account %s has status %s, not changing" % (user.user_id, user.account.status))
+        return
+
+    payload = {"status": ACTIVE}
+    try:
+        response_account = client.account[user.user_id].patch(body=payload)
+    except HTTPError, err:
+        logger.error("Account %s and UserGroup %s status were not changed", user.user_id, user.user_id)
+        raise UserStatusUpdateError("Account %s status was not changed - received HTTP code %d" % err.code)
+    else:
+        account = mkVscAccount(response_account[1])
+        if account.status == ACTIVE:
+            logger.info("Account %s status changed to %s" % (user.user_id, ACTIVE))
+        else:
+            logger.error("Account %s status was not changed", user.user_id)
+            raise UserStatusUpdateError("Account %s status was not changed, still at %s" %
+                                        (user.user_id, account.status))
+    try:
+        response_usergroup = client.account[user.user_id].usergroup.patch(body=payload)
+    except HTTPError, err:
+        logger.error("UserGroup %s status was not changed", user.user_id)
+        raise UserStatusUpdateError("UserGroup %s status was not changed - received HTTP code %d" % (user.user_id, err.code))
+
+    else:
+        usergroup = mkUserGroup(response_usergroup[1])
+        if usergroup.status == ACTIVE:
+            logger.info("UserGroup %s status changed to %s" % (user.user_id, ACTIVE))
+        else:
+            logger.error("UserGroup %s status was not changed", user.user_id)
+            raise UserStatusUpdateError("UserGroup %s status was not changed, still at %s" %
+                                        (user.user_id, usergroup.status))
+
+def process_users_quota(options, user_quota, storage_name, client):
+    """
+    Process the users' quota for the given storage.
+    """
+    error_quota = []
+    ok_quota = []
+
+    for quota in user_quota:
+        user = VscTier2AccountpageUser(quota.user, rest_client=client)
+        user.dry_run = options.dry_run
+
+        try:
+            if storage_name in ['VSC_HOME']:
+                user.set_home_quota()
+
+            if storage_name in ['VSC_DATA']:
+                user.set_data_quota()
+
+            if storage_name in ['VSC_SCRATCH_DELCATTY', 'VSC_SCRATCH_PHANPY']:
+                user.set_scratch_quota(storage_name)
+
+            ok_quota.append(quota)
+        except:
+            logger.exception("Cannot process user %s" % (user.user_id))
+            error_quota.append(quota)
+
+    return (ok_quota, error_quota)
+
+def process_users(options, account_ids, storage_name, client):
+    """
+    Process the users.
+
+    We make a distinction here between three types of filesystems.
+        - home (unique)
+            - create and populate the home directory
+        - data (unique)
+            - create the grouping fileset if needed
+            - create the user data directory
+        - scratch (multiple)
+            - create the grouping fileset if needed
+            - create the user scratch directory
+
+    """
+    error_users = []
+    ok_users = []
+
+    for vsc_id in sorted(account_ids):
+
+        user = VscTier2AccountpageUser(vsc_id, rest_client=client)
+        user.dry_run = options.dry_run
+
+        try:
+            if storage_name in ['VSC_HOME']:
+                user.create_home_dir()
+                user.set_home_quota()
+                user.populate_home_dir()
+                update_user_status(user, options, client)
+
+            if storage_name in ['VSC_DATA']:
+                user.create_data_dir()
+                user.set_data_quota()
+
+            if storage_name in ['VSC_SCRATCH_DELCATTY', 'VSC_SCRATCH_PHANPY']:
+                user.create_scratch_dir(storage_name)
+
+            ok_users.append(user)
+        except:
+            logger.exception("Cannot process user %s" % (user.user_id))
+            error_users.append(user)
+
+    return (ok_users, error_users)
